src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file src/requirements.txt src/requirements.txt
#
argparse==1.4.0
babel==2.4.0
click==6.6
coreapi==2.3.0
coreschema==0.0.4
coverage==4.2
datapunt-authorization-django==0.2.16
datapunt-authorization-levels==0.1.3
dateutils==0.6.6
debtcollector==1.13.0
django-braces==1.10.0
django-debug-toolbar==1.7
django-extensions==1.7.4
django-filter==1.0.2
django==1.11              # via datapunt-authorization-django
djangorestframework-gis==0.10.1
djangorestframework==3.6.2
django-rest-swagger==2.1.0
drf-extensions==0.3.1
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
factory-boy==2.7.0
fake-factory==0.7.2
flake8==3.3.0
iso8601==0.1.11
itypes==1.1.0
jinja2==2.9.6
keystoneauth1==2.19.0
markdown==2.6.7
markupsafe==1.0
mccabe==0.6.1
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.5
openapi-codec==1.3.1
oslo.config==3.24.0
oslo.i18n==3.15.0
oslo.serialization==2.18.0
oslo.utils==3.25.0
pbr==3.0.0
positional==1.1.1
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
pyjwt==1.5.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-keystoneclient==3.8.0
python_swiftclient==3.2.0
pytz==2017.2
pyyaml==3.12
requests==2.12.5
rfc3986==0.4.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
stevedore==1.21.0
uritemplate==3.0.0
urllib3==1.18.1
wrapt==1.10.10
xmltodict==0.10.2
